Support rdrw61x's builtin attestation key.
The rdrw61x can do attestation based on a HW key of a crypto
accellerator. For that situation it is not possible to know the key
beforehand, but for the verification of the initial attestation token
the public key needs to be exported from TF-M. This change is
implemented here.